refactor: update data types in models for improved compatibility
- Change github_id fields in PullRequest, ReviewComment, and Finding models from Integer to BigInteger for better handling of large GitHub IDs.

+++ apps/api/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, DateTime, Boolean, Text, JSON
+from sqlalchemy import Column, Integer, String, DateTime, Boolean, Text, JSON, BigInteger
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.sql import func
 from datetime import datetime
@@ -9,8 +9,8 @@ class PullRequest(Base):
     __tablename__ = "pull_requests"
     
     id = Column(Integer, primary_key=True, index=True)
-    github_id = Column(Integer, unique=True, index=True)
-    pr_number = Column(Integer)  # Changed from 'number' to 'pr_number'
+    github_id = Column(BigInteger, unique=True, index=True)  # Changed to BigInteger
+    pr_number = Column(Integer)
     title = Column(String(500))
     repo_name = Column(String(200))
     state = Column(String(50))
@@ -22,8 +22,8 @@ class ReviewComment(Base):
     __tablename__ = "review_comments"
     
     id = Column(Integer, primary_key=True, index=True)
-    github_id = Column(Integer, unique=True, index=True)
-    pr_github_id = Column(Integer, index=True)
+    github_id = Column(BigInteger, unique=True, index=True)  # Changed to BigInteger
+    pr_github_id = Column(BigInteger, index=True)  # Changed to BigInteger
     body = Column(Text)
     path = Column(String(500))
     line = Column(Integer)
@@ -34,7 +34,7 @@ class Finding(Base):
     __tablename__ = "findings"
     
     id = Column(Integer, primary_key=True, index=True)
-    pr_github_id = Column(Integer, index=True)
+    pr_github_id = Column(BigInteger, index=True)  # Changed to BigInteger
     tool = Column(String(100))
     severity = Column(String(50))
     message = Column(Text)
